move ccxt to folder
