+++ src/bitbot.py
@@ -1,6 +1,8 @@
 from PIL import Image
 import io
 from src import crypto_exchanges, stock_exchanges
+from src import stock_exchanges
+from src.crypto_exchange_server import crypto_exchanges
 from src.market_chart import MarketChart
 from src.configuration.log_decorator import info_log
 from src.chart_overlay import ChartOverlay
